requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Limiter==3.10.1
gunicorn==23.0.0
python-dotenv==1.0.1
requests==2.32.3
SpeechRecognition==3.10.0
pyttsx3==2.90
pydub==0.25.1
